PCB files
The KiCAD project has been uploaded. Includes all the files and footprint libraries required for manufacturing. 
The shield works with Arduino MEGA. 
It can also work with Arduino UNO for driving Dynamixel motors, but the additional RS-485 communication functionality won't work.
This shield can be used as a direct replacement for the official Dynamixel Arduino Shield.